Adding a method to get what Type an object is
